chore: move the old db to migrate
